requirements.txt
tweepy==4.4.0
textblob==0.15.3
nltk==3.6.5
matplotlib==3.4.3
seaborn==0.11.2
PyYAML==5.4.1
wordcloud==1.8.1
pandas==1.3.3
